Further modernization.

+++ Sources/Renderer/Handlers/StorageHandler.cpp
@@ -26,7 +26,8 @@ namespace acid
 	{
 		if (m_handlerStatus == Buffer::Status::Reset || (m_multipipeline && m_uniformBlock == nullptr) || (!m_multipipeline && m_uniformBlock != uniformBlock))
 		{
-			if ((m_size == 0 && m_uniformBlock == nullptr) || (m_uniformBlock != nullptr && m_uniformBlock != uniformBlock && m_uniformBlock->GetSize() == m_size))
+			if ((m_size == 0 && m_uniformBlock == nullptr) || (m_uniformBlock != nullptr && 
+				m_uniformBlock != uniformBlock && static_cast<uint32_t>(m_uniformBlock->GetSize()) == m_size))
 			{
 				m_size = static_cast<uint32_t>(uniformBlock->GetSize());
 			}
